chore: harden gov-pass runtime and packaging

+++ cmd/gov-pass-msi-helper/path_guard_windows.go
@@ -51,7 +51,7 @@ func safeRemoveAllWindows(path string) error {
 		return err
 	}
 	if !info.IsDir() {
-		return os.Remove(path)
+		return removeCheckedWindowsPath(path, false)
 	}
 
 	entries, err := os.ReadDir(path)
@@ -63,6 +63,31 @@ func safeRemoveAllWindows(path string) error {
 		if err := safeRemoveAllWindows(child); err != nil {
 			return err
 		}
+	}
+	return removeCheckedWindowsPath(path, true)
+}
+
+func removeCheckedWindowsPath(path string, wantDir bool) error {
+	if unsafe, err := windowsPathHasReparsePoint(path); err != nil {
+		if errors.Is(err, os.ErrNotExist) {
+			return nil
+		}
+		return err
+	} else if unsafe {
+		return fmt.Errorf("refusing to remove reparse point: %s", path)
+	}
+	info, err := os.Lstat(path)
+	if err != nil {
+		if errors.Is(err, os.ErrNotExist) {
+			return nil
+		}
+		return err
+	}
+	if wantDir && !info.IsDir() {
+		return fmt.Errorf("refusing to remove replaced non-directory: %s", path)
+	}
+	if !wantDir && info.IsDir() {
+		return fmt.Errorf("refusing to remove replaced directory as file: %s", path)
 	}
 	return os.Remove(path)
 }
